Several modifications
- Usage of python3
- '-a', '--agressive': Do not pause while downloading.
                       Default: False.
- '-n', '--dry-run':   Do not actually download anything.
                       Default: False.
- '-s', '--start':     First message in range to download.
                       Default: First available message

The default behavior is to pause 30 seconds every 1000 messages while
downloading to reduce the load on the load on the gmane servers.

The argument '-s' has replaced the usage of the number file. Also, the
default behavior has changed: If no start message is specified, all
available messages on the server will be pulled.