Optimize update function (reduce STAT calls)
Instead of checking all available messages, we start with the most
recent one and push them on a stack if we do not already have that
message. If we find one that we have, we can stop checking the rest.

Finally, we retrieve all messages on the stack and put them in the
mbox (Stack is needed to preserve the order of messages).