Silently ignore NNTPTemporaryErrors
NTPTemporaryErrors occasionally occur when an article number is not
available in a group.

Before this patch, we made several attempts to issue STAT or ARTICLE
commands for an affected message. As it turned out, usually all these
attempts fail anyway. Therefore, this patch removes the respective loops
and just skips affected messages silently (no need to confuse users
here).

As a side effect, some of the method signatures can be shortened.